Additive Shifts of Continuous-Valued Interventions based on Bounded Densities

Additive Shifts of Continuous-Valued Interventions based on Bounded Densities

Compute Ratio of Observed and Counterfactual Intervention Densities

shift_additive_bounded(
  tmle_task,
  delta,
  likelihood_base,
  max_shifted_ratio,
  fold_number,
  ...
)

shift_additive_bounded_inv(
  tmle_task,
  delta,
  likelihood_base,
  max_shifted_ratio,
  fold_number,
  ...
)

get_density_ratio(tmle_task, delta, likelihood_base, fold_number)

Arguments

tmle_task
A tmle3_Task object containing data and nodes, as described and implemented in the tmle3 package. Please refer to the documentation and supporting materials of that package for details.

delta
A numeric value giving a value of the shift to be applied to the treatment. This is an additive shift so the value is merely to be added to the observed value of the treatment node "A". In the case of the inverse additive shift, the specified value will be subtracted from the observed value of the treatment node "A".

likelihood_base
The base observed data likelihood, to be used in implementing guards that ensure that the shifted treatment does not violate the bounds induced by the support of the intervention, conditional on the covariates.

max_shifted_ratio
A numeric value indicating maximum tolerance for the ratio of the counterfactual and observed intervention densities. In particular, the shifted value of the intervention is assigned to a given observational unit when the ratio of the counterfactual intervention density to the observed intervention density is below this value.

fold_number
Whether to use cross-validated likelihood factor estimates or not. Passed directly to method get_likelihoods in tmle3.

...
Additional arguments (currently unused).

See also