Outcome Under a Grid of Shifted Interventions via Targeted Working MSM

O = (W, A, Y) W = Covariates A = Treatment (binary or categorical) Y = Outcome (binary or bounded continuous)

tmle_vimshift_msm(shift_fxn = shift_additive_bounded,
  shift_fxn_inv = shift_additive_bounded_inv, shift_grid = seq(-1, 1,
  by = 0.5), max_shifted_ratio = 2, ...)

Arguments

shift_fxn
A function defining the type of shift to be applied to the treatment. For an example, see shift_additive.

shift_fxn_inv
A function defining the inverse of the type of shift to be applied to the treatment. For an example, see shift_additive_inv.

shift_grid
A numeric vector, specification of a selection of shifts (on the level of the treatment) to be applied to the intervention. This is a value passed to the functions above for computing various values of the outcome under modulated values of the treatment.

max_shifted_ratio
A numeric value indicating the maximum tolerance for the ratio of the counterfactual and observed intervention densities. In particular, the shifted value of the intervention is assigned to a given observational unit when the ratio of counterfactual intervention density to the observed intervention density is below this value.

...
Additional arguments, passed to shift functions.